+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/2.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.0] - 2026-07-08
+## [1.1.2] - 2026-07-08
 
 ### Added
 
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - From now on, ItemManager is to be named CustomItemManager to provide more clarity on the plugin's functionality.
 - Updated the structure of the message config.
+- Updated the structure of the general config.
 
 ## [1.1.1] - 2026-07-06